redis-postgres-pipeline/pyproject.toml
[project]
name = "redis-postgres-pipeline"
version = "0.1.0"
description = "High-performance data engineering pipeline with Redis and Postgres"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "faker>=25.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
